feat(user-post-list): construct user post list with section title and section post components
