+++ src/App.js
@@ -2,6 +2,7 @@ import React from 'react';
 
 import UserPublic from './pages/user-public/index';
 import PostDetail from './pages/user-public/post-detail';
+import UserPostList from './pages/user-public/user-post-list';
 
 import './styles/global.scss';
 
@@ -10,6 +11,7 @@ const App = () => {
     <>
       <UserPublic />
       <PostDetail />
+      <UserPostList />
     </>
   );
 }
