+++ src/styles/colors.scss
@@ -1,5 +1,5 @@
 $text: #202124;
-$line: #d0d0d040;;
+$line: #d0d0d040;
 $secondary: #202124bd;
 
 .border-gray {
@@ -8,4 +8,8 @@ $secondary: #202124bd;
 
 .text-secondary {
   color: $secondary;
+}
+
+.text-tritary {
+  color: #2021245e;
 }
